chore: refactor constants
