+++ src/connectors/webwallet/constants.ts
@@ -14,3 +14,14 @@ export const DEFAULT_WEBWALLET_ICON = `<svg
       fill="currentColor"
     />
   </svg>`
+
+export const TESTNET_WHITELIST_URL =
+  "https://static.hydrogen.argent47.net/webwallet/iframe_whitelist_testnet.json"
+
+export const MAINNET_WHITELIST_URL =
+  "https://static.argent.xyz/webwallet/iframe_whitelist_mainnet.json"
+
+export const RPC_NODE_URL_TESTNET =
+  "https://api.hydrogen.argent47.net/v1/starknet/goerli/rpc/v0.5"
+export const RPC_NODE_URL_MAINNET =
+  "https://cloud.argent-api.com/v1/starknet/goerli/rpc/v0.5"
